feat(transactions): safe clone/retry, import header sync, go dep bump (#127)
* feat(frontend): demo-aware login page + README demo facts

When served on the demo host (gomoney.ft-t.dev) or with localStorage
is_demo=true, the login page skips registration, auto-fills the demo
credentials, and shows a disclaimer (open-source/self-host, no real data,
daily reset, demo-only). Addresses the Google ban on the demo instance.

README demo section updated to the new domain, daily reset, and a
do-not-enter-real-data notice; minor polish.

* chore(deps): bump x/net to v0.55.0 + migrate h2c to http.Server.Protocols

govulncheck flagged 5 vulns in golang.org/x/net/html @v0.52.0
(GO-2026-5025/5027/5028/5029/5030), all fixed in v0.55.0.

v0.55.0 deprecates golang.org/x/net/http2/h2c (SA1019). Replaced
h2c.NewHandler with the stdlib http.Server.Protocols field
(SetHTTP1 + SetUnencryptedHTTP2), preserving cleartext HTTP/2 + HTTP/1
behavior. x/net is now indirect-only.

* docs(transactions): design safe retry flow

* fix(frontend): track partial transaction saves

* feat(frontend): export transaction editor drafts

* feat(frontend): add safe transaction cloning

* fix(frontend): preserve transaction draft state

* fix(frontend): sync import review header with editor edits

The import review header rendered item.transaction while the editor built
an independent FormGroup from it, so edits to title, type or amount never
reached the header.

Emit the live draft from the editor and apply it to the reviewed item in
place, preserving object identity because the template tracks by
item.transaction. Reapply the parsed sign convention so a title edit does
not flip the amount sign.

* chore(deps): update go dependencies

Bump go-money-pb, which adds MarkTransactionsIgnored to ImportService.
Return Unimplemented for it so the handler satisfies the contract without
pulling in the feature.

Also bumps gorm, pgx, mcp-go, aws-sdk and golang.org/x packages.

* fix(rules): type mock return as int32

GetCurrencyDecimals returns int32, so gomock rejected the untyped int and
TestConvertHelpers failed before this change.

* chore: stop tracking docs/superpowers

Planning and spec notes are local working artifacts, not repo content.

* fix(frontend): bump angular to 21.2.18 for trivy CVEs

Clears 5 HIGH advisories in @angular/common and @angular/core
(CVE-2026-50170, 50171, 54266, 54267, 54268), all fixed by 21.2.17.

Framework packages peer-pin each other exactly, so they move together to
their 21.2.18 ceiling. cdk stays on its own line, which tops out at
21.2.14 and carries none of these advisories.

+++ pkg/transactions/rules/lua_helpers_test.go
@@ -96,7 +96,7 @@ func TestConvertHelpers(t *testing.T) {
 			decimal.NewFromFloat(100),
 		).Return(decimal.NewFromFloat(85.50), nil)
 
-		decimalSvc.EXPECT().GetCurrencyDecimals(gomock.Any(), "EUR").Return(2)
+		decimalSvc.EXPECT().GetCurrencyDecimals(gomock.Any(), "EUR").Return(int32(2))
 
 		interpreter := rules.NewLuaInterpreter(&rules.LuaInterpreterConfig{
 			CurrencyConverterSvc: converterSvc,
